Move cf_id_wrapper to formatters to avoid a circular dependency, and ensure that the ID is a UUID
